Some BS Part 2

+++ Assets/AndrewFolder/EnemyStateData.cs
@@ -3,9 +3,9 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "EnemyStateData", menuName = "Scriptable Objects/EnemyStateData")]
 public class EnemyStateData : ScriptableObject
 {
-    public float attackRange;
-    public float alertRange;
-    public float wanderRange;
+    public int attackRange;
+    public int alertRange;
+    public int wanderRange;
 
     public LayerMask whatIsPlayer;
 }
